Move llm_env_gen package under environments/agentic_env_gen
Pure relocation requested in PR review: the package conceptually
sits alongside the other env-construction code in
isaaclab_arena/environments/, and the directory name shifts from
llm_env_gen to agentic_env_gen to align with the broader
"agent/agentic" terminology direction (the LLM is one possible
backend; a VLM-driven agent could follow later).

Files moved:
* isaaclab_arena/llm_env_gen/{__init__,llm_agent,llm_schema,try_schema}.py
  -> isaaclab_arena/environments/agentic_env_gen/...

Updated absolute import paths in try_schema.py (CLI module path in
docstring + 2 imports) and test_llm_agent.py (docstring + 2
imports), plus comment-only path references in setup.py and
.github/workflows/ci.yml.

Strictly a directory move; class names (LLMAgent, LLMEnvSpec,
LLMResponseParseError), file names within the package, the
llm_remote_e2e pytest marker, and the test file name are all
unchanged. Those renames follow in subsequent commits to keep each
diff reviewable.
