+++ setup.py
@@ -14,9 +14,9 @@ RUNTIME_DEPS = [
     "vuer[all]",
     "lightwheel-sdk",
     "pytest",
-    # Used lazily by isaaclab_arena/llm_env_gen/* for NV_API_KEY-based LLM calls.
+    # Used lazily by isaaclab_arena/environments/agentic_env_gen/* for NV_API_KEY-based LLM calls.
     "openai",
-    # Hard dependency of isaaclab_arena/llm_env_gen/llm_schema.py (BaseModel / Field /
+    # Hard dependency of isaaclab_arena/environments/agentic_env_gen/llm_schema.py (BaseModel / Field /
     # model_validator imported at module load — not lazy).
     "pydantic>=2.0",
 ]
